+++ ecommerce-batch/src/main/java/kr/co/kimga/batch/domain/product/report/ProductStatusReport.java
@@ -13,6 +13,6 @@ public class ProductStatusReport {
     private LocalDate statDate = LocalDate.now();
 
     private String productStatus;
-    private Integer productCount;
+    private Long productCount;
     private BigDecimal avgStockQuantity;
 }
